backend/requirements.txt
fastapi
uvicorn
sqlalchemy
langgraph
langchain
langchain-openai
langchain-google-genai
langchain-huggingface
chromadb
pydantic
python-dotenv
httpx
sentence-transformers
requests
